% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{find_startstop_ORFs}
\alias{find_startstop_ORFs}
\title{Find Start-Stop ORFs}
\usage{
find_startstop_ORFs(file_name)
}
\arguments{
\item{file_name}{Path to the input FASTA file.}
}
\description{
Find Start-Stop ORFs
}
