% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{find_stopstop_ORFs}
\alias{find_stopstop_ORFs}
\title{Find Stop-Stop ORFs}
\usage{
find_stopstop_ORFs(file_name)
}
\arguments{
\item{file_name}{Path to the input FASTA file.}
}
\description{
Find Stop-Stop ORFs
}
