% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{graph_stopstop_ORFs}
\alias{graph_stopstop_ORFs}
\title{Plot Stop-Stop ORFs}
\usage{
graph_stopstop_ORFs(orf_data, output_prefix = "ORF_plot")
}
\arguments{
\item{orf_data}{Path to the input CSV file generated by find_startstop_ORFs.}

\item{output_prefix}{Character string giving the prefix for output plot files.}
}
\description{
Plot Stop-Stop ORFs
}
